pyproject.toml
[project]
name = "ai-doc-translator"
version = "0.1.0"
description = "AI-agent for translating documents (PDF, TXT, HTML) via a web interface"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "pydantic>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "alembic",
    "langchain",
    "langchain-community",
    "langgraph",
    "langchain-gigachat",
    "beautifulsoup4",
    "python-dotenv",
    "python-multipart",
    "httpx",
    "reportlab>=4.4.10",
    "langdetect>=1.0.9",
    "aiofiles>=25.1.0",
    "langchain-openai>=0.3.35",
    "pymupdf>=1.27.1",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "aiosqlite",
    "ruff",
    "black",
    "mypy",
    "types-aiofiles>=25.1.0.20251011",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
# Performance: check only app package, skip venv/cache
packages = ["app"]
exclude_gitignore = true

[[tool.mypy.overrides]]
module = "langdetect"
ignore_missing_imports = true

# langchain_core.tools @tool has no type stubs; disable untyped-decorator for tools module
[[tool.mypy.overrides]]
module = "app.agent.tools"
disable_error_code = ["untyped-decorator"]

# Skip type-checking heavy third-party libs (treated as Any)
[[tool.mypy.overrides]]
module = [
    "langchain",
    "langchain.*",
    "langchain_core",
    "langchain_core.*",
    "langchain_openai",
    "langchain_openai.*",
    "langchain_gigachat",
    "langchain_gigachat.*",
    "langgraph",
    "langgraph.*",
    "reportlab",
    "reportlab.*",
    "pymupdf",
    "pymupdf.*",
]
follow_imports = "skip"
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
